Bug: 36794
The remove icon/link issue should have been resolved.  Added support to allow removal of all attendees from the scheduler widget of a meeting/call.  Previously you could not do this because there was a minimum user check as well as a check for the assigned user to be set to be scheduled.

Modified:
modules/Meeting/jsclass_scheduler.js