Bug: 38280
Merged with 6.0.0RC3.

The problem is occurring because the string values that are being used to calculate the name display are supplied with entity encoding. So for example the " (quote) comes in as " This fools the strlen function to think that there are 6 characters instead of 1. The result is that that substr is being performed against the string value with the entity encoding. Fix is to apply the html_entity_decode function on the breadcrumb values before calculating the string lengths.

Modified:
include/utils.php
include/MVC/View/tpls/modulelistmenu.tpl