Bug: 38293
Create Group User saves are failing because the retrieveEmailAddress function in SugarEmailAddress.js (this gets compiled into sugar_grp_emails.js) is expecting buttons of type 'submit', but this module has type set to 'button'.  I relaxed the check in the method to set the savePressed boolean value to true if it is of type submit or button.

Modified:
include/SugarEmailAddress/SugarEmailAddress.js