Bug: 38393
This is a regression produced by fix for 38280. We have decided to move 38280 out to 6.2 as that problem is less critical than this issue with the multibyte display. The problem here is that these multibyte characters will become garbled with the html_entity_decode and then subsequent htmlentities (in modulelistmenu.tpl) calls.

Modified:
include/utils.php
include/MVC/Views/tpls/modulelistmenu.tpl