Bug: 38464
Issue here is that the save handling for the Users module has a lot of other javascript verifications.  The SugarEmailAddress.js code is launching a form submit if Enter key is pressed or Save button is clicked.  We need to detect if a Save button is available and, if so, simulate a click so that it can trigger the module specific javascript.  This seemed to be the minimally invasive way.

Modified:
include/SugarEmailAddress/SugarEmailAddress.js