Bug: 38569
The upgrade code to detect whether or not to remove the iFrames and Feeds directories and tables assumed this was strictly a pre 5.5.x scenario and did not check if the tables do not exist. Changed the code in post_install.php to make this check and remove the files (even custom language files in this case).

Modified (builds):
scripts_for_patch/post_install.php