Bug: 38488
Traced this issue to be a possible issue with the MySQL parser with versions 5.0.x. You can check with the query:
show variables like "%version%";

The problem appears to be that the first portion of the query is not enclosed within parentheses. Later versions of MySQL handle this properly. However, for MySQL versions 5.0.x, this causes a parser error. I was only able to reproduce using the 5.2.x stacks and then installing the Sugar 6.x versions afterwards.

Made changes to SugarBean.php to enclose the first portion of the query with parentheses.

Modified:
data/SugarBean.php