Bug: 38620
For MSSQL, the call to drop the table on a table that does not exist gives errors and results in log calls using an instance of LoggerManager that does not have the setLogger method loaded yet. Changed the post_install script to check if the table exists before attempting to drop.

Modified (build):
scripts_for_patch/post_install.php