Bug: 38765
Made changes to modules/Campaigns/utils.php for Enterprise editions so that when the get_subscription_lists_query and subscribe functions are called, we check if the current user has the portal_only flag set.  If so, disable the team security clause checks.  This fixes the problem of both the newsletters not appearing and the subscribe/unsubscribe functionality not working.

Modified:
modules/Campaigns/utils.php