Bug: 39099
Updated the post_install.php file so that we also check to see if the custom/Extension/application/Ext/Include files exists for the Feeds and iFames module when removing. If so, remove these as well. Also added check to remove these modules if the modules/[Feeds/iFrames] directories do not exist. Finally, updated the manifest.php for the module loadable Feeds module so that we do not install images since there is no images directory bundled with the module.

Modified (build):
scripts_for_patch/post_install.php

Modified (sugarmods):
feeds/manifest.php