Merging with maint_6_0_2
Bug: 38859

Removed the code in TeamSetLink.php save method where the teamList variable would be set to a blank array in the event that the in_workflow flag was set. This is no longer needed as the lines of code after that handle the case where the team_id is not set in the teamList variable.

Modified:
moduldes/Teams/TeamSetLink.php

Added:
tests/Bug38859
tests/Bug32738