Bug: 39157
Error seems to be a result of creating the column alias in SugarWidgetReportField.php. There's a 28 character limit for the alias because some database implementations have a limit on this alias length. In this case, the unique aliases were truncated because the length of the alias (table name + column name) exceeded 28 characters. We are altering the code in SugarWidgetReportField.php (_get_column_alias method) so that we make a better attempt to ensure that the aliases are unique by running an md5 on the alias and taking the first 4 values from this to append to the first 24 characters.

Modified:
include/generic/SugarWidgets/SugarWidgetReportField.php