Bug: 33075
This seems to occur on select browsers (able to reproduce on Chrome, but not IE and Firefox). Need to add safety handling for the item name and item value textboxes so that a return does not trigger a form submission. To attempt to reproduce, try the mentioned steps to reproduce using Chrome on Windows.

Modified:
modules/ModuleBuilder/tpls/MBModule/dropdown.tpl