Bug: 38909
As a first pass, changed the deletion of user to also mark the private team deleted column to be true (1).  Fixed issue with showing the deleted user in the "Assign To" dropdown.  The issue of full team reassignment mentioned in this thread is contingent on a configuration setting in the config.php (enable_team_module_save) which is not enabled by default and not exposed via the admin interface.  We may need some discussion around this to see if we should enable it by default and expose it through admin interface.  With that setting enabled, we can direct the user to another page after the user reassignment section is completed to reassign other team records that the user may have been a part of (via teamsets).

Modified:
modules/Users/controller.php
modules/Users/reassignUserRecords.php