Bug: 36950
This is a problem with the query using the same join table alias for both the billing and shipping accounts.  In this scenario where both billing and shipping account names need to be displayed, we need another set of join queries on a different quotes_accounts table alias.  Changed the table alias in the Quotes vardefs.php for the shipping_account_name entry to allow for this.

Modified:
sugarcrm/modules/Quotes/vardefs.php