Bug: 40043
This error is caused by the javascript engine for the Firefox 3.0.15 version running on the stack environment.  It turns out that in style.js, the attribute children in nodes.children is not defined.  Instead, we need to use the YAHOO.util.Dom libraries to help us properly query the elements and also add a safety check to ensure that it is an object (looking for a div element) so that the style property value can be determined.  This fix is in the onSubmenuShow function.

Modified:
sugarcrm/themes/Sugar/js/style.js