Bug: 38519
The jssource/src_files directory was removed sometime around 5.5.3 and is no longer shipped or used in the product.  To test this, you will have to go through the process of upgrading from < 5.5.3 all the way to 6.2.  After the upgrade to 6.2, the jssource/src_files directory should no longer exist.  Added code in the post_install.php file to check if this directory exists and remove all its contents if so.

Modified:
build/scripts_for_patch/post_install.php