Bug: 38969
This bug stems from the process_union_list_query method in SugarBean.php.  As we iterate through the fields, we are setting the subpanel's bean value only if the value evaluates to true using the !empty check.  However, in the case of the integer zero value, !empty is still evaluating to true.  We could change the check to:

!empty($row[$field]) || (isset($row[$field]) && $row[$field] == 0)

but this brings some overhead.  The isset($row[$field]) check should be suffice.

Modified:
sugarcrm/data/SugarBean.php