Bug: 40455
This appears to be caused by an error from the json encoding when the calcField value for the Sugar.portal object declaration is blank.  You will see a message "Syntax error };\n" in the browser console.  Traced the error and fixed the layout template to handle empty calculated field value for this json entry.

Modified:
sugarcrm/modules/ModuleBuilder/tpls/layoutview.tpl