Bug # 40373 (Last Viewed for the Home module.)
Added a new variable $tracker_visibilty (default value true) and a new function get_tracker_visibility() in SugarBean, which returns the value of the variable tracker_visibility, whose value can be true or false. This value is used in track_view() function of SugarBean.

Have added a new variable $tracjer_visibilty in Quota.php. its value is set to false, as this module is not supposed to be tracked.