Bug: 40433
Traced this to a problem with Report.php and the mssql driver.  In the create_query method, the currency fields would have 2 select entries.  As a result, when the $field_list_name_array variable was rewritten, one portion of the currency (either currency id or currency value) would be lost.  Made changes to accommodate for the behavior of currency fields.

Modified:
sugarcrm/modules/Reports/Report.php