39161: Disabled the contact field from edit/detail/quickcreate layouts in studio
The contact field in calls is not a real field in the DB, and is only used to display a primary contact on the list view. On the detail view, the contacts subpanel shows all contacts related to the call. On the Edit View, the invitees functionality is how contacts should be added to the call.