Bug 36845
Made changes to support global searches for custom modules.  Also added support for upgrades so that we
scan for custom modules and add the unified_search attribute to the vardefs file where there is a
SearchFields.php file defined.  Added unit test as well to simulate scanning directories and upgrading the
vardefs.php files.