Bug: 40876
It appears that changes made to the vardefs for team_security template and the basic modules not set the team_count and id field to be required.  From PopupSmarty.php the handleSave call to AccountFormBase.php was setting the $useRequired variable to true.  As a result, this method would just return null when the !checkRequired function was run since there are no id or team_count values passed in at this point when creating a new Account.  Made changes to PopupSmarty.php to set useRequired parameter to false in the event that no id is found in the $_REQUEST array.