Bug: 40412
Found issues with some browsers that were interpreting the type value for the phone field to be "tel".  Firefox was more forgiving and returning type of "text" which is what the code in sugar_3.js is searching for.  Modified the sugar field editview template so that we set the type to "text" instead of "tel".