Bug: 40129
Fixed assumption of the installation process when merging config_si.php settings.  Previously, it was assumed that a config.php file exists so that it could properly merge the settings.  This worked for upgrades or in cases where config.php exists.  The new fix will run the merge after the config.php file has been created.