Bug 40178: Initial Support in SugarLogic for related fields.
Right now link fields are not understood directly and strings must be passed to the RelatedFieldExpression instead. Values do populate correctly however and can be used successfully in formulas. Next step is for link fields to be proper types in both the javascript and php versions of SugarLogic for validation, loading, and multiple relationship traversing purposes.