Bug: 41120
Note for the latest comments from Jennifer 2/9/2011, the test scenario is not the same as that listed in the description.  To reproduce this issue and see the popup, you can do the following:
1) In Your own User EditView, enable External Accounts for Google Docs
2) Go to Documents->Create New Document
3) Change type to "Google Docs"
4) In the "File on External Source" widget that now appears, click on the select button to bring up the popup

Made changes to fix the formatting issues as well as remove the direct_url variable from schema table and code.