Bug: 42261
The problem is that the history files that were being saved were using some datetime format like "2011-02-15T20:11:35+00:00".  This blows up on Windows because of the special characters that the Windows filesystem does not accept for extensions.  There was also an issue on all systems where clicking on "View History" button on Studio caused errors.  Changed the code in History.php to save files with a timestamp extension instead.  This is what the view.history.php code was expecting as well (when clicking on "View History").