bug 42259 (Error Encountered When Trying to Send to Multiple Recipients with Commas in Name)
comma in the name was getting replaced by ::;::, so the name was being treated as an email address, and hence it was failing the pattern matching.

so now if comma is used for separating email addresses then only it will be replaced by ::;::  and if it is used in any name e.g. smith, jr comma will not be replaced by ::;::