Bug: 42321
I looked into this issue more today and am concluding that this is an issue with the Zend_Gdata_Doc class and the Google Document 1.0 API implementation.  There are certain file types that are not returned in the result set when UPLOADED to Google Docs (.png, .tpl, etc.).  However, files that are CREATED on Google Docs will appear.  This limitation needs to be noted.  I have added some notes to the ExtAPIGoogle.php file as well.