Bug: 42383
Took a while to trace this issue down.  It seems there was some errant code that was not removed in modules/Campaigns/Charts.php.  The problem was that a fetch call was made to the result set and then a subsequent fetch call was made that would result in the code always returning null (note the test environment was SQL Server).  The code did not make sense because right after the $camp_data variable was set it was immediately declared again as an empty array.