Bug: 42561
Applied fix for scenario where 1000s separator is set to empty.  Fix unformatNumberNoParse method in sugar_3.js so that we skip running the regular expression to replace all global occurrences of the 1000s separator.  The decimal issue may need to be resolved via a feature improvement to allow for non-empty values.