Bug: 42617
The skipped query  "ALTER TABLE email_addr_bean_rel ALTER COLUMN bean_module varchar(100) NULL" is most likely caused by the fact that the alter statement on the bean_module column is linked to indexes in the table.  The alter statement is executed via the .sql file.  The new repair code can now correctly amend this problem so I have taken it out of the .sql file.  Tested the upgrade without the manual SQL statements in the .sql file and table columns are altered correctly.