Bug: 42739
This error was occurring because the date and number field widgets for range searches were defaulting to use the "equals" value.  However, the dropdown value associated with Equals label is "=".  As a result, the widget would think that the value was a macro value (Next Month, Last Week, etc.) and not show a text input field.