Bug: 42691
The equals operation (=) against a datetime field will not work because of the time portions.  We are setting the search field type to datetime or datetimecombo in the event that an equals operation (=) search is done against such fields.  This allows for the search code to use a between search against two values to adjust for the time range.  Same issue was also found with float, currency and decimal fields so I've applied fixes to that as well.