Bug: 42378
Search fixes were resolved through changes for 42691.  However, this bug exposed another issue with SavedSearch module.  That is, we serialize the user's saved search values as is, but fail to convert date, datetime or datetimecombo type field values to db format.  This will cause problems when user changes their date format and then reloads a saved search previously saved with a different date format.  So I also made changes to SavedSearch class so that when saving search contents, date fields will be serialized in db format before it is serialized to the contents column.  Then upon populating the $_REQUEST and $_GET variables in the populateRequest method, we convert these db format date fields to the user's date format preferences.  This resolves the issue whereby user preferences for date format are changed between saved searches.  Add unit test as well.