Bug: 42790
It seems we are not checking an empty value against a field that is not required.  In this case, even though the password expiration login attempts field was not selected via radio button, the addToValidate call was still made.  As a result, the blank field is not an integer and the validation fails.  The error was not appearing because disabling Added fix to sugar_3.js.