Bug: 42877
The file object type sugar object is a bit different from the Documents module.  For one, we do not have the revisions storage with custom file type sugar objects.  Also, the support for external files as we have in the Documents module is not replicated in 6.2.  For now, we will set the filename field to be hidden from Studio since this value is set internally in the File.php code.  There is no need to expose this field in studio.  We resolve the duplicate "File Name" entries this way.