Bug: 43134
The upgrade is okay because the custom code block for the filename field is no longer needed (it was removed in the 6.2 metadata file) and this customCode block was the OOTB code prior to 6.2.  The javascript error is specific to IE and how the object attributes were not properly terminated (a trailing comma for the last element).  Other browsers are more forgiving, but this is a common error for IE.  Made the fix to the javascript file for File SugarField.