Bug: 43145
Fixed Oracle warning message about inconsistent types.  This error seems to occur from the query in SugarEmailAddress.php, getAddressByGUID method

"SELECT ea.*, ear.* FROM email_addresses ea
                LEFT JOIN email_addr_bean_rel ear ON ea.id = ear.email_address_id
                WHERE ear.bean_module = '{$module}'
                AND ear.bean_id = '{$id}'
                AND ear.deleted = 0
                ORDER BY ear.reply_to_address, ear.primary_address DESC"

This may be a PHP / Oracle driver bug and it seems to occur intermittently.  I have made changes to the SQL to be explicit in which columns to query.  With those changes, I have not seen the error occur.