Bug: 43157
For silent upgrades, we incorrectly rebuild the cached language files for the modules because we do not correctly clear out the cached language files.  The cache/modules/Contacts/language/en_us.lang.php file that was being rebuilt from silent upgrades did not correctly retrieve the template objects and thus the teams labels were missing (from modules/Teams/language/en_us.lang.php).  As a result, the LBL_TEAMS used would be the one found in include/language/en_us.lang.php (this had "Team:" and not the expected "Teams:" found in the Teams module).  I have moved the call to deleteCache in silentUpgrade_step2.php to occur before the calls to rebuild the vardefs/database.