Bug: 43420
Looked into this a little more.  To circumvent the user formatting of the date values, we will use the global flag disable_date_format in MassUpdate.php so that all calls to retrieve will not apply the user date formatting.  This way when the Call or Meeting is saved from MassUpdate, the exected call to create a datetime instance from database format will succeed.