Bug: 43616
Error occurs because we should have never permitted such actions of dragging items in tabs to subpanels or subpanels to tabs.  The same issue occurs for Connectors search properties in that the elements from one module may be moved to another module.  This also should not be permitted.  Added grouping properties to the lists so that drag an drop operations remain between the allowed groups.  You should not be able to move tab items to subpanels and vice versa.  For Connectors, you may not drag search fields from one module to another module.