Bug: 43653
Fixed issue where, when no user_preferences are for global search are found, we use the unified_modules_search_display.php file (what has been chosen) and not the unified_modules_search.php file (what is permissible by default).  In this case, since Tasks module is not enabled by default for global search, the code was displaying this in the user preferences tabs because it used unified_modules_search.php.