Bug: 44030
Fixed the SugarSpot.php code for modules that do not have unified_search set on a module level to not unset email searches.  What happens is that the search field definition gets passed into the search form code and then the join clause is executed.  The search was not being made because SugarSpot.php was unsetting the search field since there is no 'email' vardefs entry.