Bug: 44136
Upgrade errors appears to stem from not being able to locate the upgrade zip file in the uploads directory.  Removed the code to unzip and re-copy the upgrade zip file in silentUpgrade_step2.php as this is already done in silentUpgrade_step1.php.   Added some additional logging messages to better support error detection as well.