Bug: 43395
Made changes to the quick search formatting of results so that we do not process the salutation and title fields when performing a quicksearch.  Also changed TemplateHandler to not use getContactQS method to build quicksearch code for Contacts search (this should be handled instead by the generic query function).  Also changed popup displays to not invoke the locale formatting for names and salutations as well.  Added unit test Bug43395Test.php.