Bug: 44374
Looks like this is a result of bug 38519 where the src_files directory was completely removed to get rid of old javascript libraries.  What should have occurred is perhaps a series of specific checks against the offending sub directories instead of getting rid of the entire directory.  Made the changes to remove the offending directories:

- 'jssource/src_files/include/javascript/ext-2.0'
- 'jssource/src_files/include/javascript/ext-1.1.1'
- 'jssource/src_files/include/javascript/yui'