Bug: 44154
After looking at this issue, it appears that certain email service providers impose this restriction.  The code in InboundEmail.php (handleCreateCase method) is setting the value correctly:

$reply->from_addr = $fromAddress;

I think at this point, we should include some information on the field so that user is aware of this possible limitation.

I've added help text to the "From Address" field in Admin->Inbound Email records for the english language.  Assigned to ProductManagement to review the text.  It says:

"The email address specified here may not appear in the from address section of the email sent due to restrictions on the mail service provider.  In these circumstances, the email address defined in the outgoing mail server will be used."