Bug: 34993
Fixed handling of default values for custom fields.  When a custom field is created with a default value, the default value
should not be set in the existing _cstm table entries.  This is because it would create inconsistent behavior with all records
that don't have a corresponding _cstm entry (their default value would not be set).  The default value specified in a custom field
should mean that is the value that will be elected when editing a record (if none has been selected).