Bug: 44836
I added quickcreatedefs.php file for the module which duplicates the editviewdefs.php file.  The other option to fix this given the current code base is to not have the subpanel quick create display an inline form, but take the user to a new full EditView form.  I went with the quickcreatedefs.php file for now.