Bug: 45092
This was a regression caused by fix for Bug 44496.  The assumption here was that numCols would always be defined, but looking through the code in charts and dashlets, the numCols variable is typically defined in PRO only blocks.  This ia a safety check we need here to default to two columns in the event that numCols is not even defined.