Bug: 40658
Added support to show links for relationships in subpanels.  Also show links for custom relate fields.
Changed SubpanelMetaDataParser so that we apply the necessary updates to subpanel definition file to show the link.
Changed the DynamicField.php file so that the relate table query includes the relate table record id.
Added unit test.