Bug: 46191
Added code to also update all of the include/database directory before the SugarCache code is copied over.  The problem was that the cache code was updated, but the database code was not.  As a result, the code was expecting the database to return no cache, but the database code was still attempting to return a failed cache entry (fooling system to think the user did not exist in the system).