Bug 38864 - Added ability to customize the modules available for the Reports module as well as those that should be exempt from the Reports module.
This is done by moving the $exemptModules array out of config.php and into reportmoduledefs.php, as well as defining a new array $additionalModules for a list of new modules to add. Also cleaned up most of the logic here to improve the overall speed of the code. Unit test should help explain everything.