Bug: 47555
Fixed the get_field_name method in TemplateField.php.  The name it was returning was appending an extra '_c' suffix when the custom field
already had the '_c' suffix.  The result was that the SearchFields.php definition were incorrectly created and thus the search was never
properly run against the custom field.