Bug #47650 - SugarBean::get_list modification
this modifies the SugarBean::get_list method to support passing in an array of fields so that you can just query on the fields that you need instead of running the potentialy hugh query that needs to be run to pull back a list of Beans.

Also this modifies SugarBean::updateDependantField() to not run when it's running a WebService call as they are not needed since it's setting display flags and the like.