Bug 47723 'Summation with details report causes SQL errors on MSSQL'
The issue is that when creating a summation report with details the $groupby variable for query type 'query' might end up looking like '(leads.status' which results in ORDER BY CharIndex((leads.status + ..etc. when using MSSQL. The issue was already fixed for query type 'summary_query' and this fix just reapplies it to the 'query' query type.

@contrib apextwo