Bug: 47512
Traced this to the __destruct method in SugarTheme.php.  What seems to be happening is that the possibility for SugarTheme destruct to be called when creating custom fields.  The SugarTheme destruct method was making an attempt to call sugar_mkdir to create the cache sugar theme directory as a safety.  However, in the course of being destructed, the cache directory cannot be created anyway and thus a log error in include/utils/sugar_file_utils sugar_mkdir method is attempted.  The log instance here is null as well since this was also probably destroyed.  Fix was to not call sugar_mkdir. The previous version of the code prior to changes were not making the call to sugar_mkdir anyway.  We could do a mkdir perhaps, but I don't think this is necessary since it's happening in the destruct method anyway.