Bug: 47864
Made changes to the form base classes.  Created a new PersonFormBase class which LeadFormBase extends.  This allows for better reusability and testing of behavior between ContactFormBase and LeadFormBase code.  The code between the two classes are very similar and this hopefully opens up the possibility to further re-factor some of the other legacy code in there.  Added two simple initial unit tests as well.