Bug: 48550
Revert fix for 48486.  The fix for 48486 resulted in two reports_to_id hidden fields to be loaded onto the user's edit view screen.  As a result, the validation function failed because the validation would incorrectly report that the second reports_to_id field was empty (value was set in the first one only).