Bug: 48571
Add fix to SugarTheme.php to resolve issues with setting the default theme.  The logic to use array_pop now
results in the RTL theme being the default theme in cases where the global default theme is not set in configuration files and no user preference theme is defined.  This is because the RTL/rtl key is last in the list of keys when scanning the directory of themes.