Bug: 48879
Changed the global search quick view to not display the field values for fields which the user does not have ACL field access to.  An interesting thing is obviously in this case the account name itself is shown in the search results which brings up a point of discussion for the future- do we just not show search results for the accounts in this case?  I also optimized some of the code in the global search quick view and also added php documentation to the ACLField.php file.