Bug: 48898
For some web servers, the register value for the user_array was still cached after creating a new user.  As a result,
the lookup to newly created users was not working and we could not set a default assigned user name (for the newly created user).
Fixed to call clear_register_value when creating a new User from modules/Users/Save.php.