Bug: 41296
Added support to match duplicate Contact without email requirement.  If a Contact to sync does not have an email address and the first and last name matches another Contact in the system that also does not have an email address, then the first matched Contact on server is considered the duplicate record.