Bug: 48889
Fixed the create_new_list_query method in Employees.php.  This was a regression caused by 47813 where the return_array function was set to true and did not respect the value sent in from the web service calls (false).  As a result, the query was erroneous and bombed out.  Added unit test tests/service/Bug48889Test.php