Bug: 49006
Fixed errors with subpanels when the user subpanel tabs option is enabled.  The problems are a result of the Ajax UI changes and the timing of loading of javascript files as well as changes to the way in using a global $currentModule variable and a global app bean are referenced.  Fixed a bunch of javascript errors, added extra checks to guard against errors and fixed problem with referencing relationships when the emails subpanels are called (which failed in modules/SNIP/utils.php function).  Added a SODA test in tests/dev/Bug49006.xml.