Bug: 48607
This bug is interesting because what seems to have happened is that the yyDDList.js file in the most recent versions had the ygLogger call commented out.  This file hasn't been changed in over 1.5 years now and probably longer.  What likely happened is that the upgrade path on this instance skipped the version that would have brought in the new yyDDList.js file (with the ygLogger call commented out).  So this error occurs even before the upgrade.  I am updating the yyDDList.js file to generate an md5 change and get this file to be brought into the upgrade zip files so that the new version can be copied over during upgrades.