Bug: 49188
Fixed the convert lead screen.  Revert to just using images instead of sprites for the edit and delete buttons.  This is because the way the
layout currently works, it is using javavscript innerHTML to draw out the images.  The problem with the sprites code is that it inserts all these other tags that are not properly enclosed resulting in the javascript breaking.