Bug: 48879
Made feature change for field level access control.  If the user does not have field access permission to the field(s) that are used to display the global search results, then we instead show a link that will run the search to simulate the "Show All" button and take the user to the list view results there.

Also reformatted some of the code and added type checks to not run searches for fields that don't match expected search criteria.