Bug: 49852
The fix to 48726 resolves the javascript error (window.document.forms[form_name].elements[the_key] ).  However, I found another fix to resolve the sorting suite failures and possibly some other SODA tests.  The problem was that we are prompting for the confirmation dialog whenever popup data returns for new records.  This should not be happening as there is no need to confirm a dialog for changing data that was empty or was exactly the same as before.  The reason for this was because of popup_helper.js, the oldData variable was calculated by adding a space between the label and the data whereas the label variable did not include a space.  Also, the logic check newData != label will always be true since they are constructed differently.  This should have been newData != oldData.