Bug: 49898
Added modification to SoapSugarUsers.php as a backdoor workaround to current plugin limitations.  The current implementations
of the plugin code use the old soap.php API where search_by_module accepts a username and password.  This workaround allows
a session id to be passed in for validation as well.

Added unit test.