Bug: 49996
Remove the reliance on the site_url parameter in the header.tpl file.  Here, there was probably an incorrect specification to not include the port numbers.  However, since there is no connection check against the portal instance URL, we assume that the site_url is the correct URL including port number.