Bug: 49873
Second attempt to resolve this problem with deleteing Documents, but having these Documents (related to Contracts) still appear in the Contracts subpanel.  There are two paths to delete a Document from the user interface

1) ListView - Calls MassUpdate and then mark_deleted on Document instance
2) DetailView - Calls modules/Documents/Delete.php or custom one (if found)

The first option was not using the Delete.php file that contained the logic to remove relationships.  Moved the calls to mark_relationshps_deleted in Document.php to resolve this.  This function now handles removing the revisions and contracts relationships.

Added unit test.