Bug: 32797
Fix the problem in VardefManager applyGlobalAccountRequirements method where the assumption was that any field named 'account_name' would be strictly required if the system configuration specified so.  This fix resolves the issue of not being able to set the required setting of Targets and Leads back to non-required from studio.