Bug: 50046
We've had some back and forth with this issue.  This was originally addressed by bug 45571 (https://github.com/sugarcrm/Mango/pull/1379), but was later superceded by https://github.com/sugarcrm/Mango/pull/1639.  I believe the recursive utf8 encoding is still necessary since the changes made to Safe.php do not necessarily ensure the content is utf8 encoded, but rather that it is not repacked with UTF7 encoding.

The fix here calls the recursive encoding on the output object to ensure all elements are utf8 encoded.