Bug: 50220
Fix the tracker_session insertion problem in DB2.  This was due to the incorrect assumption that a query
to the date columns always resulted in a datetime format string.  This is not the case for DB2.  Added
defensive guards in TrackerSessionDatabaseStore.php.  Also fixed SugarApplication and populateSeedData files
to correct convert the values.

Added unit test.