Bug: 48938
Call encodeURIComponent from the grid.js code, don't force all paramsToUrl calls to use encodeURIComponent.
The problem with the previous commit from https://github.com/sugarcrm/Mango/pull/1669 was that it would force
all the calls to paramsToUrl to be encoded and thus causing some double encoding in portions of the application.

If we want a combination method we could perhaps create a encodeParamsToUrl function instead.