Bug: 50318
This was a regression caused by 47645 in which it was assumed all email address fields have a parent span id.  This was not the case for the users view because of the way the element is loaded (not through a smarty function from a .tpl file , but directly from UserViewHelper.php).