Bug: 50617
Added member variable to mark whether or not mark_deleted should truly be run when cascade_delete against a Workflow instance is called.
From the Save operation where an existing workflow just has a target module (base_module) modified, we do not truly mark the Workflow instance
to be deleted.