Bug: 49385
Version 5 of the fix.  This version correctly uses the secondary_select support that was persent in ListViewData.php.  Code changes done so as to have minimal impact on other areas.  Only list view displays for Meetings and Calls will trigger the singleSelect variable to be false when calling create_new_list_query (which then triggers the secondary_select support).  Fixed the secondary_select code that was broken (the list of ids created was incorrect).  Added unit test.