Fix bug #50695 - include the right files and add a few extra checks.
This bug fixes two smaller bugs within. Firstly, the javascript parser now
checks to see if there isn't a trailing new line at the end of the line
for a comment. This can happen when the comment is the last thing in the
file and the file does not contain a newline character at the end of it.
It also defines $primedInput as an array so that there is no error
thrown when accessing an element of the array.

This bug also fixes the require order for some of the functions
(minify.php).