Bug: 51004
In working with Riva more, we have found some other changes to the registry.php file so that we can preserve the association keys on SOAP calls via PHP.  This solution, tested with them, resolves the issue we had with the WSDL not loading.  Tested the WSDL generated on SOAPUI, Chrome Extension Plug-in and Unit tests.  Riva has also verified.  I have to re-open this again, sorry, but wanted to let you know in case you needed to re-verify as well.