Bug: 51311
Fixed the error with the massageFieldDef function in MssqlManager.  The problem here is that for the longtext type defined in vardefs.php files, the dbType would be set to nvarchar(max) and the checks in massageFieldDef was checking for text or ntext only.  This should have just used isTextType method so we fixed the isTextType method to also look for longtext and nvarchar(max) dbTypes and return true if so.