Move `standalone in nav` to separate file...[read more]
So I found a cleaner code of this in https://gist.github.com/1811212,
In previous commit I correct this function in mobile.js. But in this commit
as instruction in js itself comments say that it should be first script
in <head />. So I am adding it to heads of all our .htmls here. Feel
free to revert this commit if you don't want it yet to live in head
since we cannot test it anyway on webkit as standalone app.