Fix #51575: Portal install fails on Enterprise.
The issue here was that we were sending back an array of user
properties. However, in SugarBean.php, we call a method of the User
class on the array, causing the SOAP call from the portal to fail. This
patch ensures that the object retrieved is a SugarBean and not an array
containing values from the database row.